requirements.txt
flask==3.0.0
tensorflow-cpu==2.16.1
scikit-learn==1.4.0
numpy==1.26.4
joblib==1.3.2
gunicorn==21.2.0
